Make popout accept children
